联调成功

+++ iot_manager/src/main/java/com/example/iot_manager/vo/SimpleModelVO.java
@@ -13,7 +13,7 @@ public class SimpleModelVO {
 
   private String name;
 
-  private int num;
+  private double num;
